+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Added microservice utilities for easier testing of provided samples.
 
+* Added Tenant Options API support.
+
 ## Version 1.4
 
 * Fixed https://github.com/SoftwareAG/cumulocity-python-api/issues/25